+++ R/nano.R
@@ -194,9 +194,9 @@ nano <- function(protocol = c("bus", "pair", "push", "pull", "pub", "sub",
   nano[["stat"]] <- function(name) stat(socket, name = name)
 
   switch(attr(socket, "protocol"),
-         req =,
+         req = ,
          rep = {
-           nano[["context_open"]] <- function(verify = TRUE) {
+           nano[["context_open"]] <- function() {
              if (is.null(sock2)) sock2 <<- socket
              nano[["context_close"]] <- function() if (length(sock2)) {
                r <- close(socket)
@@ -205,11 +205,11 @@ nano <- function(protocol = c("bus", "pair", "push", "pull", "pub", "sub",
                rm(list = c("context", "context_close"), envir = nano)
                r
              }
-            socket <<- nano[["context"]] <- context(sock2, verify = verify)
+            socket <<- nano[["context"]] <- context(sock2)
            }
          },
          sub = {
-           nano[["context_open"]] <- function(verify = TRUE) {
+           nano[["context_open"]] <- function() {
              if (is.null(sock2)) sock2 <<- socket
              nano[["context_close"]] <- function() if (length(sock2)) {
                r <- close(socket)
@@ -218,7 +218,7 @@ nano <- function(protocol = c("bus", "pair", "push", "pull", "pub", "sub",
                rm(list = c("context", "context_close"), envir = nano)
                r
              }
-             socket <<- nano[["context"]] <- context(sock2, verify = verify)
+             socket <<- nano[["context"]] <- context(sock2)
            }
            nano[["subscribe"]] <- function(topic = NULL)
              subscribe(socket, topic = topic)
@@ -226,7 +226,7 @@ nano <- function(protocol = c("bus", "pair", "push", "pull", "pub", "sub",
              unsubscribe(socket, topic = topic)
          },
          surveyor = {
-           nano[["context_open"]] <- function(verify = TRUE) {
+           nano[["context_open"]] <- function() {
              if (is.null(sock2)) sock2 <<- socket
              nano[["context_close"]] <- function() if (length(sock2)) {
                r <- close(socket)
@@ -235,13 +235,13 @@ nano <- function(protocol = c("bus", "pair", "push", "pull", "pub", "sub",
                rm(list = c("context", "context_close"), envir = nano)
                r
              }
-             socket <<- nano[["context"]] <- context(sock2, verify = verify)
+             socket <<- nano[["context"]] <- context(sock2)
            }
            nano[["survey_time"]] <- function(value = 1000L)
              survey_time(socket, value = value)
          },
          respondent = {
-           nano[["context_open"]] <- function(verify = TRUE) {
+           nano[["context_open"]] <- function() {
              if (is.null(sock2)) sock2 <<- socket
              nano[["context_close"]] <- function() if (length(sock2)) {
                r <- close(socket)
@@ -250,7 +250,7 @@ nano <- function(protocol = c("bus", "pair", "push", "pull", "pub", "sub",
                rm(list = c("context", "context_close"), envir = nano)
                r
              }
-             socket <<- nano[["context"]] <- context(sock2, verify = verify)
+             socket <<- nano[["context"]] <- context(sock2)
            }
          },
          NULL)
@@ -376,7 +376,7 @@ print.ncurlAio <- function(x, ...) {
 print.ncurlSession <- function(x, ...) {
 
   cat(sprintf("< ncurlSession >\n - %s\n",
-              if (length(attr(x, "aio"))) "use transact() to return data" else "not active" ), file = stdout())
+              if (length(attr(x, "aio"))) "use transact() to return data" else "not active"), file = stdout())
   invisible(x)
 
 }
@@ -404,6 +404,15 @@ print.errorValue <- function(x, ...) {
 print.conditionVariable <- function(x, ...) {
 
   cat("< conditionVariable >\n", file = stdout())
+  invisible(x)
+
+}
+
+#' @export
+#'
+print.thread <- function(x, ...) {
+
+  cat("< thread >\n", file = stdout())
   invisible(x)
 
 }
@@ -466,4 +475,3 @@ print.conditionVariable <- function(x, ...) {
 #'
 .DollarNames.ncurlAio <- function(x, pattern = "")
   grep(pattern, c("status", "headers", "raw", "data"), value = TRUE, fixed = TRUE)
-