+++ NEWS.md
@@ -1,3 +1,21 @@
+# nanonext 0.8.3
+
+#### New Features
+
+* Implements `cv_signal()` and `timed_signal()` for signalling a condition variable, the latter after a specified time (from a newly-created thread).
+* Implements `.context()`, a performance alternative to `context()` that does not create the full object.
+* Adds utility `nanonext_version()` for providing the package version, NNG and mbed TLS library versions in a single string.
+* `ncurl()` gains a 'timeout' argument.
+
+#### Updates
+
+* Removes 'verify' argument of `context()` (changed to '...' for compatibility) as `request()` and `request_signal()` have been rendered safe internally for use with timeouts.
+* The name of the single argument to `msleep()` has been changed to 'time' from 'msec'.
+* Functions `pipe_notify()`, `lock()` and `unlock()` now error if unsuccessful rather than returning with a warning.
+* For compiling bundled 'libmbedtls' and 'libnng' libraries from source, R's configured C compiler is now chosen over the system default where this is different.
+* Bundled 'libnng' source updated to v1.6.0 alpha (c5e9d8a).
+* Bundled 'libmbedtls' source updated to v3.4.0.
+
 # nanonext 0.8.2
 
 #### New Features